% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_doc.R
\docType{data}
\name{theme_def_ex}
\alias{theme_def_ex}
\title{Example theme_def}
\format{
A data.frame with 2 columns (ELEMENT, COLOR) and 38 rows corresponding to the
portion of the theme it modifies.
}
\usage{
theme_def_ex
}
\description{
If defining your own theme_def, it's in the exact specification provided by this
example. If it is not, rs.ui will not run.
}
\keyword{datasets}
