% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tint_shade.R
\name{tint.this}
\alias{tint.this}
\alias{shade.this}
\title{Tints or Shades a Color}
\usage{
tint.this("#7cb7a3", .15)

shade.this("#7cb7a3", .20)
}
\arguments{
\item{color}{A valid color specification as either a hex code or name.}

\item{value}{A value between 0 and 1. This is the percentage the supplied color
is shifted by.}
}
\description{
Tints or Shades a Color
}
\examples{
rs.ui(main_color = tint.this("#3d2b56"))
rs.ui(main_color = shade.this("#ae2573", .25))
}
